engine: move RPC method registry into nasty-engine, delete nasty-apidoc
Co-locates the docs metadata with the dispatcher so they stop drifting.
Adds `nasty-engine --dump-docs <dir>` as the new CLI for generating
docs/api.md, replacing the standalone nasty-apidoc binary.

Bug fixes from drift between the old apidoc registry and the live router:
- subvolume.list/.get/.attach/.detach docs claimed `pool`, runtime parses
  `filesystem` (router/subvolume.rs require_str("filesystem"))
- snapshot.list: same `pool` → `filesystem` rename
- auth.token.create: docs claimed `pool`, runtime struct has `filesystem`
- Session and ApiToken docs now reflect the engine's real types — the old
  apidoc mirrors were missing `client_ip`, `created_at`, `owner`,
  `must_change_password`, `allowed_ips`

Schema fidelity: 9 share/iscsi/nvmeof Delete/Remove methods now use their
actual Request structs (already existed in nasty-sharing — apidoc had
used inline Literal strings); the other 28 inline-parsed methods now
emit proper JSON Schema objects via MethodParams::AdHoc.

This is PR1a of a 4-PR series toward in-engine OpenAPI docs and a
vendored Swagger UI at /api/docs:
  - PR1a (this PR): registry move, 106 currently-documented methods
  - PR1b: backfill 150 methods the dispatcher accepts but apidoc never
    documented (vm.*, apps.*, backup.*, smb.user.*, system.hardware.*,
    system.secure_boot.*, system.acme.*, system.tls.*, etc.)
  - PR2: OpenAPI 3.1 emitter + REST gateway at /api/v1/{path}
  - PR3: vendor swagger-ui-dist via flake.nix, serve /api/docs
