+++ engine/Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "2"
 members = [
     "nasty-engine",
-    "nasty-apidoc",
     "nasty-common",
     "nasty-metrics",
     "nasty-storage",
