--- engine/nasty-apidoc/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "nasty-apidoc"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[[bin]]
-name = "nasty-apidoc"
-path = "src/main.rs"
-
-[dependencies]
-nasty-storage = { path = "../nasty-storage" }
-nasty-sharing = { path = "../nasty-sharing" }
-nasty-system = { path = "../nasty-system" }
-schemars.workspace = true
-serde.workspace = true
-serde_json.workspace = true
